[HAX] restore_history_point(): skip steps where wininfo wasn't updated
!! Tagged as HAX because the way this was implemented is rather
inelegant. The proper way to do this requires significantly more
changes. !!

Added a new field to wininfo to indicate whether the grid was updated.

Added a new var wininfo_undo_history_cursor to keep track of the last
saved wininfo that was updated as part of a grid update.

Update restore_history_point() to check for whether that field is set to
1 before deciding to restore to restore to that particular wininfo.

Update save_history_point() to update wininfo_undo_history_cursor only
if the wininfo was updated as part of a grid update.